% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{autoxgboost-package}
\alias{autoxgboost}
\alias{autoxgboost-package}
\title{autoxgboost: Automatic tuning and fitting of xgboost}
\description{
Automatic tuning and fitting of 'xgboost'. Emulates autoxgboost for mlr3.
}
\author{
\strong{Maintainer}: Martin Binder \email{mlr.developer@mb706.com}

}
